+++ src/ExtractXliff/Program.cs
@@ -1,9 +1,8 @@
-// Copyright (c) 2017 SIL International
+// Copyright (c) 2020 SIL International
 // This software is licensed under the MIT License (http://opensource.org/licenses/MIT)
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using L10NSharp;
 using L10NSharp.CodeReader;